Fix takeDamage() to work with armourBonus from player inventory
- Remove getBaseHealth(), getBaseArmour(), getBaseDamage()
- Superseded with getCurrent______()

- Remove healthBonus from Item- same functionality as armourBonus

- Add Item tests

+++ Room.cpp
@@ -86,7 +86,7 @@ void Room::addItem(Item* item)
 
 Boss* Room::getBoss(int index)
 {
-    if (index >= 0 && index < bosses.size())
+    if (index >= 0 && index < static_cast<int>(bosses.size()))
     {
         return bosses[index];
     }
@@ -95,7 +95,7 @@ Boss* Room::getBoss(int index)
 
 Item* Room::getItem(int index)
 {
-    if (index >= 0 && index < items.size())
+    if (index >= 0 && index < static_cast<int>(items.size()))
     {
         return items[index];
     }
